First real run.
Tokens are assigned as before, but in cases of mismatches,
we go and look if there is the same token id in the old traceability
tier, in which case we take start and stop arguments from this element.

In this case we sucessfully propagate Darinka's edits forward, except for the
overlapping speakers, which will be handled manually for the 6 cases we have that.